requirements.txt
asgiref==3.2.10
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decouple==0.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
dj-stripe==2.3.0
Django==3.0.8
django-allauth==0.43.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-paypal==1.0.0
django-rest-framework==0.1.0
django-richtextfield==1.6
djangorestframework==3.12.2
docutils==0.16
gunicorn==20.0.4
idna==2.9
jsonfield==3.1.0
Kivy-Garden==0.1.4
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
urllib3==1.25.9
whitenoise==5.2.0
